easyocr/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
easyocr[cpu]
opencv-python-headless
numpy
Pillow==10.1.0 